requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
entsoe-py>=0.6.0
python-dotenv>=1.0.0
pyarrow>=14.0.0
statsmodels>=0.14.0
